// types.d.ts
export type PaymentEntry = {
    id: number;
    created: number;
    updated: number;
    monetary_account_id: number;
    amount: number;
    currency: string;
    description: string;
    type: string;
    iban: string;
    name: string;
    category_code?: string;
    subtype: string;
    balance_after: number;
};